Remove perceval/backends/__init__.py
When installing in the system (eg, via a pip package), this
file overwrites the one in Perceval, if this package is installed
by chance after Perceval. Right now that is not a problem,
because both have the same content. But I think it is better to
avoid that potential problem, by just removing this file,
which is not needed anyway.