[init] Add perceval/__init__.py as in perceval
When installing this package after perceval, its (formerly
empty) perceval/__init__.py file overwrites (or supersedes,
depending on how it is installed) the one in the perceval
package. That causes errors: see for example grimoirelab/perceval#190
or grimoirelab/perceval#171.

This is not the best fix, since it requires maintaining
this file in sync in all perceval and perceval-* packackes,
but for now, it does the trick.