Add support to run tests using setup.py file
To run user-defined commands using setup.py it is necessary
to add Command derivated classes. In this case, TestCommand
class was included to run tests.

The inclusion of `pkg_resources` is necessary to ensure that
Perceval is imported from the right path, inserting its location
in the first position of `sys.path` variable. This change is due
to the use of namespaces by Perceval.

To avoid import problems, it was necessary to add `base.py`
script, an empty `__init__.py` file into tests directory and
change how the package is imported in `test_functest.py`.