Include metadata files and tests in sdist packages
Source distribution packages will start including
some missing metadata files such as the license and
the list of authors.

It also will include the code needed to test the
software. This is useful for distros which could
need to pass some tests before generating the packages
(e.g, in rpm or deb formats).

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>